+++ torch_frame/data/dataset.py
@@ -116,6 +116,17 @@ class Dataset(ABC):
             cols.remove(self.target_col)
         return cols
 
+    @property
+    @requires_post_materialization
+    def num_classes(self) -> int:
+        if StatType.COUNT not in self.col_stats[self.target_col]:
+            raise ValueError(
+                f"num_classes attribute is only supported when the target "
+                f"column ({self.target_col}) stats contains StatType.COUNT, "
+                f"but only the following target column stats are calculated: "
+                f"{list(self.col_stats[self.target_col].keys())}.")
+        return len(self.col_stats[self.target_col][StatType.COUNT][0])
+
     # Materialization #########################################################
 
     def materialize(self, device: Optional[torch.device] = None) -> 'Dataset':
@@ -248,3 +259,19 @@ class Dataset(ABC):
         dataset.col_to_stype = {col: self.col_to_stype[col] for col in cols}
 
         return dataset
+
+    def get_split_dataset(self, split: str) -> 'Dataset':
+        r"""Get splitted dataset defined in the 'split' column of
+        :obj:`self.df`.
+
+        Args:
+            split (str): The split name. Should be 'train', 'val', or 'test'.
+        """
+        if split not in ['train', 'val', 'test']:
+            raise ValueError(f"The split named {split} is not available. "
+                             f"Needs to either 'train', 'val', or 'test'.")
+        if 'split' not in self.df:
+            raise ValueError(
+                f"'get_split_dataset' is not supported for {self}.")
+        indices = self.df.index[self.df['split'] == split].tolist()
+        return self[indices]