+++ torch_frame/datasets/__init__.py
@@ -6,6 +6,9 @@ from .fake import FakeDataset
 from .forest_cover_type import ForestCoverType
 from .mushroom import Mushroom
 from .poker_hand import PokerHand
+from .bank_marketing import BankMarketing
+from .tabular_benchmark import TabularBenchmark
+from .yandex import Yandex
 
 real_world_datasets = [
     'Titanic',
@@ -13,6 +16,9 @@ real_world_datasets = [
     'ForestCoverType',
     'Mushroom',
     'PokerHand',
+    'BankMarketing',
+    'TabularBenchmark',
+    'Yandex',
 ]
 
 synthetic_datasets = [