+++ torch_frame/nn/conv/trompt_conv.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn.functional as F
 from torch import Tensor
-from torch.nn import GroupNorm, Linear, Parameter
+from torch.nn import GroupNorm, LayerNorm, Linear, Parameter
 
 from torch_frame.nn.conv import TableConv
 
@@ -37,14 +37,19 @@ class TromptConv(TableConv):
         # Group norm
         self.group_norm = GroupNorm(num_groups=num_groups,
                                     num_channels=num_prompts)
+        # Layer norms
+        self.layer_norm_e_column = LayerNorm(channels)
+        self.layer_norm_e_prompt = LayerNorm(channels)
         self.reset_parameters()
 
     def reset_parameters(self):
-        torch.nn.init.normal_(self.embedding_column, std=0.1)
-        torch.nn.init.normal_(self.embedding_prompt, std=0.1)
+        torch.nn.init.normal_(self.embedding_column, std=0.01)
+        torch.nn.init.normal_(self.embedding_prompt, std=0.01)
         self.lin.reset_parameters()
-        torch.nn.init.normal_(self.weight, std=0.1)
+        torch.nn.init.normal_(self.weight, std=0.01)
         self.group_norm.reset_parameters()
+        self.layer_norm_e_column.reset_parameters()
+        self.layer_norm_e_prompt.reset_parameters()
 
     def forward(self, x: Tensor, x_prompt: Tensor) -> Tensor:
         r"""Transforms :obj:`x` and :obj:`x_prompt` into :obj:`x_prompt` for
@@ -66,7 +71,8 @@ class TromptConv(TableConv):
 
         # Step1: Compute stacked_e_prompt
         # [batch_size, num_prompts, channels]
-        stacked_e_prompt = self.embedding_prompt.repeat(batch_size, 1, 1)
+        embedding_prompt = self.layer_norm_e_prompt(self.embedding_prompt)
+        stacked_e_prompt = embedding_prompt.repeat(batch_size, 1, 1)
         # [batch_size, num_prompts, 2*channels]
         stacked_e_prompt_cat = torch.cat([stacked_e_prompt, x_prompt], dim=-1)
         # [batch_size, num_prompts, channels]
@@ -75,7 +81,8 @@ class TromptConv(TableConv):
 
         # Step2: Compute stacked_e_column
         # [batch_size, num_cols, channels]
-        stacked_e_column = self.embedding_column.repeat(batch_size, 1, 1)
+        embedding_column = self.layer_norm_e_column(self.embedding_column)
+        stacked_e_column = embedding_column.repeat(batch_size, 1, 1)
 
         # Step 3: Use stacked_e_prompt and stacked_e_column to compute
         # M_importance
@@ -89,9 +96,13 @@ class TromptConv(TableConv):
 
         # Step 4: Expand x ([batch_size, num_cols, channels]) into
         # shape [batch_size, num_prompts, num_cols, channels]
-        x = torch.einsum('ijl,k->ikjl', x, self.weight)
-        x = F.relu(x)
-        x = self.group_norm(x)
+        z = torch.einsum('ijl,k->ikjl', x, self.weight)
+        z = F.relu(z)
+        # [batch_size, num_cols, channels] ->
+        # [batch_size, num_prompts, num_cols, channels]
+        x = x.unsqueeze(1).repeat(1, self.num_prompts, 1, 1)
+        # Residual connection
+        x = self.group_norm(z) + x
 
         # Step 5: Use x and m_importance to compute x_prompt for the next layer
         # x * m_importance: shape [batch_size, num_prompts, num_cols, channels]