+++ torch_frame/nn/encoder/stype_encoder.py
@@ -1,13 +1,24 @@
+import math
 from abc import ABC, abstractmethod
 from typing import Any, Dict, List, Optional, Set
 
 import torch
 from torch import Tensor
-from torch.nn import Embedding, ModuleList, Parameter
+from torch.nn import Embedding, ModuleList, Parameter, Sequential
+from torch.nn.init import kaiming_uniform_
 
 from torch_frame import stype
 from torch_frame.data.stats import StatType
 from torch_frame.nn.base import Module
+
+from ..utils.init import attenuated_kaiming_uniform_
+
+
+def reset_parameters_soft(module: Module):
+    r"""Call reset_parameters() only when it exists. Skip activation module."""
+    if (hasattr(module, 'reset_parameters')
+            and callable(module.reset_parameters)):
+        module.reset_parameters()
 
 
 class StypeEncoder(Module, ABC):
@@ -19,6 +30,10 @@ class StypeEncoder(Module, ABC):
         out_channels (int): The output channel dimensionality
         stats_list (List[Dict[StatType, Any]]): The list of stats for each
             column within the same stype.
+        post_module (Module, optional): The posthoc module applied to the
+            output, such as activation function and normalization. Must
+            preserve the shape of the output. If :obj:`None`, no module will be
+            applied to the output. (default: :obj:`None`)
     """
     supported_stypes: Set[stype] = {}
     LAZY_ATTRS = {'out_channels', 'stats_list'}
@@ -28,8 +43,9 @@ class StypeEncoder(Module, ABC):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     @abstractmethod
     def forward(self, x: Tensor):
@@ -37,7 +53,27 @@ class StypeEncoder(Module, ABC):
 
     @abstractmethod
     def reset_parameters(self):
-        raise NotImplementedError
+        r"""Initialize the parameters of `post_module`"""
+        if self.post_module is not None:
+            if isinstance(self.post_module, Sequential):
+                for m in self.post_module:
+                    reset_parameters_soft(m)
+            else:
+                reset_parameters_soft(self.post_module)
+
+    def post_forward(self, out: Tensor) -> Tensor:
+        r"""Post-forward function applied to :obj:`out` of shape
+        [batch_size, num_cols, channels]. It also returns :obj:`out` of the
+        same shape."""
+        if self.post_module is not None:
+            shape_before = out.shape
+            out = self.post_module(out)
+            if out.shape != shape_before:
+                raise RuntimeError(
+                    f"post_module must not alter the shape of the tensor, but "
+                    f"it changed the shape from {shape_before} to "
+                    f"{out.shape}.")
+        return out
 
 
 class EmbeddingEncoder(StypeEncoder):
@@ -50,8 +86,9 @@ class EmbeddingEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         self.embs = ModuleList([])
@@ -65,6 +102,11 @@ class EmbeddingEncoder(StypeEncoder):
                     padding_idx=0,
                 ))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        for emb in self.embs:
+            emb.reset_parameters()
 
     def forward(self, x: Tensor):
         r"""Maps input :obj:`x` from TensorFrame (shape [batch_size, num_cols])
@@ -82,12 +124,8 @@ class EmbeddingEncoder(StypeEncoder):
         for i, emb in enumerate(self.embs):
             xs.append(emb(x[:, i]))
         # [batch_size, num_cols, hidden_channels]
-        x = torch.stack(xs, dim=1)
-        return x
-
-    def reset_parameters(self):
-        for emb in self.embs:
-            emb.reset_parameters()
+        out = torch.stack(xs, dim=1)
+        return self.post_forward(out)
 
 
 class LinearEncoder(StypeEncoder):
@@ -101,8 +139,9 @@ class LinearEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         mean = torch.tensor(
@@ -115,6 +154,11 @@ class LinearEncoder(StypeEncoder):
         self.weight = Parameter(torch.empty(num_cols, self.out_channels))
         self.bias = Parameter(torch.empty(num_cols, self.out_channels))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        torch.nn.init.normal_(self.weight, std=0.01)
+        torch.nn.init.zeros_(self.bias)
 
     def forward(self, x: Tensor):
         r"""Maps input :obj:`x` from TensorFrame (shape [batch_size, num_cols])
@@ -129,33 +173,23 @@ class LinearEncoder(StypeEncoder):
         # [batch_size, num_cols, channels] + [num_cols, channels]
         # -> [batch_size, num_cols, channels]
         x = x_lin + self.bias
-        return torch.nan_to_num(x, nan=0)
-
-    def reset_parameters(self):
-        torch.nn.init.normal_(self.weight, std=0.1)
-        torch.nn.init.zeros_(self.bias)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
 
 class LinearBucketEncoder(StypeEncoder):
     r"""A numerical converter that transforms a tensor into a piecewise
     linear representation, followed by a linear transformation. The original
-    encoding is described in https://arxiv.org/abs/2203.05556.
-
-    Args:
-        out_channels (int): The output channel dimensionality
-        stats_list (List[Dict[StatType, Any]]): The list of stats for each
-            column within the same stype.
-            - StatType.QUANTILES: The min, 25th, 50th, 75th quantile, and max
-            of the column.
-    """
+    encoding is described in https://arxiv.org/abs/2203.05556"""
     supported_stypes = {stype.numerical}
 
     def __init__(
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
     ):
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         # The min, 25th, 50th, 75th quantile, and max of the column.
@@ -167,6 +201,12 @@ class LinearBucketEncoder(StypeEncoder):
             torch.empty(num_cols, self.interval.shape[-1], self.out_channels))
         self.bias = Parameter(torch.empty(num_cols, self.out_channels))
         self.reset_parameters()
+
+    def reset_parameters(self):
+        super().reset_parameters()
+        # Reset learnable parameters of the linear transformation
+        torch.nn.init.normal_(self.weight, std=0.01)
+        torch.nn.init.zeros_(self.bias)
 
     def forward(self, x: Tensor):
         encoded_values = []
@@ -196,12 +236,8 @@ class LinearBucketEncoder(StypeEncoder):
         # -> [batch_size, num_cols, channels]
         x_lin = torch.einsum('ijk,jkl->ijl', out, self.weight)
         x = x_lin + self.bias
-        return torch.nan_to_num(x, nan=0)
-
-    def reset_parameters(self):
-        # Reset learnable parameters of the linear transformation
-        torch.nn.init.normal_(self.weight, std=0.1)
-        torch.nn.init.zeros_(self.bias)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
 
 class LinearPeriodicEncoder(StypeEncoder):
@@ -212,7 +248,6 @@ class LinearPeriodicEncoder(StypeEncoder):
     in https://arxiv.org/abs/2203.05556.
 
     Args:
-        out_channels (int): The output channel dimensionality
         n_bins (int): Number of bins for periodic encoding
     """
     supported_stypes = {stype.numerical}
@@ -221,10 +256,11 @@ class LinearPeriodicEncoder(StypeEncoder):
         self,
         out_channels: Optional[int] = None,
         stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
         n_bins: Optional[int] = 16,
     ):
         self.n_bins = n_bins
-        super().__init__(out_channels, stats_list)
+        super().__init__(out_channels, stats_list, post_module)
 
     def init_modules(self):
         mean = torch.tensor(
@@ -239,6 +275,11 @@ class LinearPeriodicEncoder(StypeEncoder):
             torch.empty((num_cols, self.n_bins * 2, self.out_channels)))
         self.reset_parameters()
 
+    def reset_parameters(self):
+        super().reset_parameters()
+        torch.nn.init.normal_(self.linear_in, std=0.01)
+        torch.nn.init.normal_(self.linear_out, std=0.01)
+
     def forward(self, x: Tensor):
         x = (x - self.mean) / self.std
         # Compute the value 'v' by scaling the input 'x' with
@@ -252,9 +293,68 @@ class LinearPeriodicEncoder(StypeEncoder):
         # [batch_size, num_cols, num_buckets],[num_cols, num_buckets, channels]
         # -> [batch_size, num_cols, channels]
         x = torch.einsum('ijk,jkl->ijl', x, self.linear_out)
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
-        return x
+
+class ExcelFormerEncoder(StypeEncoder):
+    r""" An attention based encoder that transforms input numerical features
+    to a 3-dimentional tensor.
+    Before being fed to the embedding layer, numerical features are normalized
+    and categorical features are transformed into numerical features by the
+    CatBoost Encoder implemented with the Sklearn Python package. The features
+    are then ranked based on mutural information.
+    The original encoding is described in https://arxiv.org/pdf/2301.02819
+
+    Args:
+        out_channels (int): The output channel dimensionality
+        stats_list (List[Dict[StatType, Any]]): The list of stats for each
+            column within the same stype.
+    """
+    supported_stypes = {stype.numerical}
+
+    def __init__(
+        self,
+        out_channels: int,
+        stats_list: Optional[List[Dict[StatType, Any]]] = None,
+        post_module: Optional[Module] = None,
+    ):
+        super().__init__(out_channels, stats_list, post_module)
+
+    def init_modules(self):
+        mean = torch.tensor(
+            [stats[StatType.MEAN] for stats in self.stats_list])
+        self.register_buffer('mean', mean)
+        std = torch.tensor([stats[StatType.STD]
+                            for stats in self.stats_list]) + 1e-6
+        self.register_buffer('std', std)
+        num_cols = len(self.stats_list)
+        self.W_1 = Parameter(Tensor(num_cols, self.out_channels))
+        self.W_2 = Parameter(Tensor(num_cols, self.out_channels))
+        self.b_1 = Parameter(Tensor(num_cols, self.out_channels))
+        self.b_2 = Parameter(Tensor(num_cols, self.out_channels))
+        self.reset_parameters()
+
+    def forward(self, x: Tensor) -> Tensor:
+        r"""Transforming :obj:`x` into output embeddings.
+
+        Args:
+            x (Tensor): Input column-wise tensor of shape
+                [batch_size, num_cols]
+
+        Returns:
+            x (TensorFrame): [batch_size, num_cols, out_channels].
+        """
+        x = (x - self.mean) / self.std
+        x1 = self.W_1[None] * x[:, :, None] + self.b_1[None]
+        x2 = self.W_2[None] * x[:, :, None] + self.b_2[None]
+        x = torch.tanh(x1) * x2
+        out = torch.nan_to_num(x, nan=0)
+        return self.post_forward(out)
 
     def reset_parameters(self):
-        torch.nn.init.normal_(self.linear_in, std=0.1)
-        torch.nn.init.normal_(self.linear_out, std=0.1)
+        super().reset_parameters()
+        attenuated_kaiming_uniform_(self.W_1)
+        attenuated_kaiming_uniform_(self.W_2)
+        kaiming_uniform_(self.b_1, a=math.sqrt(5))
+        kaiming_uniform_(self.b_2, a=math.sqrt(5))