+++ CHANGELOG.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Added `ExcelFormerEncoder` ([#42](https://github.com/pyg-team/pytorch-frame/pull/42))
+- Make `FTTransformer` take tensor_frame as input ([#45](https://github.com/pyg-team/pytorch-frame/pull/45))
+- Added `Tompt` example ([#39](https://github.com/pyg-team/pytorch-frame/pull/39))
+- Added `post_module` in `StypeEncoder` ([#43](https://github.com/pyg-team/pytorch-frame/pull/43))
+- Added `FTTransformer` ([#40](https://github.com/pyg-team/pytorch-frame/pull/40), [#41](https://github.com/pyg-team/pytorch-frame/pull/41))
+- Added `ExcelFormer` ([#26](https://github.com/pyg-team/pytorch-frame/pull/26))
+- Added `Yandex` collections ([#37](https://github.com/pyg-team/pytorch-frame/pull/37))
+- Added `TabularBenchmark` collections ([#33](https://github.com/pyg-team/pytorch-frame/pull/33))
+- Added the `Bank Marketing` dataset ([#34](https://github.com/pyg-team/pytorch-frame/pull/34))
 - Added the `Mushroom`, `Forest Cover Type`, and `Poker Hand` datasets ([#32](https://github.com/pyg-team/pytorch-frame/pull/32))
 - Added `PeriodicEncoder` ([#31](https://github.com/pyg-team/pytorch-frame/pull/31))
 - Added `NaN` handling in `StypeEncoder` ([#28](https://github.com/pyg-team/pytorch-frame/pull/28))
@@ -18,7 +27,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added indexing/shuffling and column select functionality in `Dataset` ([#18](https://github.com/pyg-team/pytorch-frame/pull/18), [#19](https://github.com/pyg-team/pytorch-frame/pull/19))
 - Added the `Adult Census Income` dataset ([#17](https://github.com/pyg-team/pytorch-frame/pull/17))
 - Added column-level statistics and dataset materialization ([#15](https://github.com/pyg-team/pytorch-frame/pull/15))
-- Added `FTTransformer` ([#12](https://github.com/pyg-team/pytorch-frame/pull/12))
+- Added `FTTransformerConvs` ([#12](https://github.com/pyg-team/pytorch-frame/pull/12))
 - Added documentation ([#13](https://github.com/pyg-team/pytorch-frame/pull/13))
 - Added `DataLoader` capabilities ([#11](https://github.com/pyg-team/pytorch-frame/pull/11))
 - Added `TensorFrame.index_select` functionality ([#10](https://github.com/pyg-team/pytorch-frame/pull/10))